Introduce new -unsigned packages to house the pre-built .efi binaries
We will eventually change grub-install to only require these packages
(or the signed ones where available) instead of the -bin ones. In this
first step, we just add the new packages, and transitional depends from
the -bin packages to them, such that the functionality of -bin packages
remains the same. Loong64 is an outlier here, having not shipped a grub
before, we do not have a transitional depends.

We also need to have Breaks and Replaces but only for the packages
that we built monolithic binaries in before, aka the amd64, ia32,
and arm64 targets.

The next step is to patch grub-install and then we can change the
dependencies of the grub-efi-{amd64,...} packages to point to the
-signed binaries on the signed platforms, and the -unsigned ones
on the unsigned ones.

Fixup